+++ SimpleCalculator/Application/AppDelegate.swift
