+++ SimpleCalculator/Application/Assets.xcassets/AppIcon.appiconset/Contents.json
