+++ SimpleCalculator/Application/Assets.xcassets/Contents.json
