+++ SimpleCalculator/Application/Base.lproj/LaunchScreen.storyboard
