+++ SimpleCalculator/Application/Info.plist
